Login Credentials and Database Backup
Admin login credentials can be viewed through the text file.

There are two ways to import the database of this Wordpress site.

One is using the *.sql file by uploading it through the phpmyadmin control panel.

Another way is to turn off all services (apache and mysql), copy the mysql folder, and paste it to /xampp folder. Turn on all services back, and the Wordpress site should load correctly.